added submit button to each field type form
