+++ src/models/__init__.py
@@ -1,7 +1,8 @@
-from .base_db_model import (Base, create_table, db, get_all_tables_information,
-                            remove_table, update_table_content)
+from .base_db_model import Base, db
 from .column import ColumnInfo
 from .content import Content
+from .model_update import (create_table, get_all_tables_information,
+                           remove_table, update_table_content)
 from .web_user import WebUser
 
 __all__ =[
