+++ src/modules/content_manager/static/css/pop-up/content-manager-pop-up.css
@@ -16,7 +16,7 @@
     height: 3.5rem;
     padding: 0 2rem;
     border-radius: 0.5rem;
-    margin: 1rem 2.5rem;
+    margin: 2rem 2.5rem;
 }
 
 #full-description-text {
